feat: Implement socket handlers for channel management, desktop agent, and electron integration
- Added channel handlers for managing channel changes, receiver hello messages, and intent resolution.
- Introduced desktop agent handlers for session setup, directory listing, app launch registration, and client state updates.
- Created disconnect handlers to manage app and channel disconnections effectively.
- Implemented electron handlers for app discovery and desktop agent initialization.
- Refactored socket service initialization to register all handlers in a modular way.
- Established types and utility functions for socket connection management and state reporting.
