+++ packages/socket/src/desktop-agent/handlers/index.ts
@@ -1,0 +1,9 @@
+// Export all handler types and utilities
+export * from './types';
+
+// Export individual handlers
+export * from './electron.handlers';
+export * from './desktop-agent.handlers';
+export * from './app.handlers';
+export * from './channel.handlers';
+export * from './disconnect.handlers';
